separated UI file from main file
